cargo clippy fix

+++ rust/src/index/vector/kmeans.rs
@@ -52,13 +52,13 @@ pub async fn train_kmeans(
     mut rng: impl Rng,
 ) -> Result<Float32Array> {
     let num_rows = array.len() / dimension;
-    if num_rows < k as usize {
+    if num_rows < k {
         return Err(crate::Error::Index(format!(
             "KMeans: can not train {k} centroids with {num_rows} vectors, choose a smaller K (< {num_rows}) instead"
         )));
     }
     // Ony sample 256 * num_clusters. See Faiss
-    let data = if num_rows > 256 * k as usize {
+    let data = if num_rows > 256 * k {
         println!(
             "Sample {} out of {} to train kmeans of {} dim, {} clusters",
             256 * k,
@@ -66,7 +66,7 @@ pub async fn train_kmeans(
             dimension,
             k,
         );
-        let sample_size = 256 * k as usize;
+        let sample_size = 256 * k;
         let chosen = (0..num_rows).choose_multiple(&mut rng, sample_size);
         let mut builder = Float32Builder::with_capacity(sample_size * dimension);
         for idx in chosen.iter() {
